Move to action defs to avoid circular import
