+++ src/ansible_navigator/action_defs.py
@@ -1,0 +1,25 @@
+"""Pre-defined objects used by actions."""
+
+
+from typing import NamedTuple
+from typing import Union
+
+
+class RunReturn(NamedTuple):
+    """The base return object for any action's run method."""
+
+    #: A message to display immediately before exiting
+    message: str
+    #: The return code for the for the process
+    return_code: int
+
+
+class RunInteractiveReturn(RunReturn):
+    """The return object for an actions' run interactive method."""
+
+
+class RunStdoutReturn(RunReturn):
+    """The return object for an actions' run stdout method."""
+
+
+ActionReturn = Union[RunReturn, RunInteractiveReturn, RunStdoutReturn]
