+++ src/ansible_navigator/actions/builder.py
@@ -10,7 +10,7 @@ from typing import Optional
 from typing import Tuple
 
 from ..action_base import ActionBase
-from ..action_base import RunStdoutReturn
+from ..action_defs import RunStdoutReturn
 from ..configuration_subsystem import ApplicationConfiguration
 from ..configuration_subsystem.definitions import Constants
 from ..runner import Command
