+++ src/ansible_navigator/actions/config.py
@@ -15,7 +15,7 @@ from typing import Union
 from .._yaml import Loader
 from .._yaml import yaml
 from ..action_base import ActionBase
-from ..action_base import RunStdoutReturn
+from ..action_defs import RunStdoutReturn
 from ..app_public import AppPublic
 from ..configuration_subsystem import ApplicationConfiguration
 from ..runner import AnsibleConfig
