+++ tests/integration/_action_run_test.py
@@ -11,7 +11,7 @@ from typing import List
 from typing import Optional
 from typing import Tuple
 
-from ansible_navigator.action_base import RunStdoutReturn
+from ansible_navigator.action_defs import RunStdoutReturn
 from ansible_navigator.app_public import AppPublic
 from ansible_navigator.configuration_subsystem import Constants as C
 from ansible_navigator.configuration_subsystem import NavigatorConfiguration
